Fix Integer to Long Add test Should get by Id

+++ registry/src/main/java/com/gl/registry/controller/DeviceController.java
@@ -26,7 +26,7 @@ public class DeviceController {
     }
 
     @GetMapping("/{id}")
-    public DeviceResponseVO getById(@PathVariable Integer id) {
+    public DeviceResponseVO getById(@PathVariable Long id) {
         return deviceService.getById(id);
     }
